requirements.txt

pymongo==4.14.1
robotframework==7.3.2
robotframework-appiumlibrary==3.1

